General update function description
The updates folder will contain the update handlers. That can be used
to update, create savely (new) documents. So the user don't have to
take care about revision numbers, even if there are concurrent
requests. The function will always get two parameters (`doc` and
`request`). While the `doc` will be the one that is called via
`/_updates/[:_id of the docuemnt]`. The update function should
**always** return an array, e.g. `[ new saved document, response]`.