Added general model description
Models will and should be used to handle the data that is needed for a
view. So you should always have one model for one view. Models will
contain at least one property `props`. This will be the object where
all the properties should be defined.

Derived properties (also known as computed properties) are properties
of the state object that depend on other properties (from `props`,
`session`, or even `derived`) to determine their value.