added testing to the repository
An example server-side test is created. The general usage in development is to execute one test file directly.

But to ensure that all devs remember to execute all test before they commit a pre-commit hook is also added. Every dev has to exec once the command `npm install` in the root folder. It will install the hook into the `.git/hooks` folder.